pressure: force non-zero if pending
When calculating the percentage of the batch size for pending items on the queue, because of precision loss, the result can be zero, which can instigate an infinite loop during ->flush.  Force pressure to have a minimum value of '1' if there are items pending on the queue to prevent this.